test: moved pages under yjs path
