+++ tests/e2e/src/pages/yjs/node-redis/storage/index.tsx
@@ -1,8 +1,8 @@
 import { NextPage } from "next";
 import { useRouter } from "next/router";
-import React, { useState } from "react";
-import { CloudflarePresenceRoom } from "../../../components";
-import { PluvRoomProvider } from "../../../pluv-io/yjs/cloudflare";
+import { useState } from "react";
+import { NodeRedisStorageRoom } from "../../../../components";
+import { PluvRoomProvider } from "../../../../pluv-io/yjs/node-redis";
 
 export const Page: NextPage = () => {
     const router = useRouter();
@@ -11,7 +11,7 @@ export const Page: NextPage = () => {
 
     if (!router.isReady) return null;
 
-    const roomId = (router.query.room as string) ?? "e2e-cloudflare-presence";
+    const roomId = (router.query.room as string) ?? "e2e-node-redis-storage";
 
     return (
         <div>
@@ -37,13 +37,13 @@ export const Page: NextPage = () => {
             {enabled && (
                 <PluvRoomProvider
                     debug={{
-                        input: ["$GET_OTHERS", "SEND_MESSAGE"],
-                        output: ["$REGISTERED"],
+                        input: ["$UPDATE_STORAGE"],
+                        output: ["$STORAGE_UPDATED", "$STORAGE_RECEIVED"],
                     }}
                     initialPresence={{ count: 0 }}
                     room={roomId}
                 >
-                    <CloudflarePresenceRoom />
+                    <NodeRedisStorageRoom />
                 </PluvRoomProvider>
             )}
         </div>
