+++ tests/e2e/src/pages/yjs/node/presence/index.tsx
@@ -1,8 +1,8 @@
 import { NextPage } from "next";
 import { useRouter } from "next/router";
 import React, { useState } from "react";
-import { NodePresenceRoom } from "../../../components";
-import { PluvRoomProvider } from "../../../pluv-io/yjs/node";
+import { NodePresenceRoom } from "../../../../components";
+import { PluvRoomProvider } from "../../../../pluv-io/yjs/node";
 
 export const Page: NextPage = () => {
     const router = useRouter();
