+++ tests/e2e/src/pluv-io/loro/node.ts
@@ -2,14 +2,14 @@ import { addonIndexedDB } from "@pluv/addon-indexeddb";
 import { yjs } from "@pluv/crdt-yjs";
 import { createBundle, createClient } from "@pluv/react";
 import { z } from "zod";
-import type { io } from "../server/node";
+import type { io } from "../../server/yjs/node";
 
 const client = createClient<typeof io>({
     authEndpoint: (room) => {
-        return `http://localhost:3102/api/pluv/authorize?room=${room}`;
+        return `http://localhost:3112/api/pluv/authorize?room=${room}`;
     },
     wsEndpoint: (room) => {
-        return `ws://localhost:3102/api/pluv/room/${room}`;
+        return `ws://localhost:3112/api/pluv/room/${room}`;
     },
 });
 
@@ -44,11 +44,6 @@ export const {
     usePluvTransact,
     usePluvUndo,
 } = createRoomBundle({
-    addons: [
-        addonIndexedDB({
-            enabled: (room) => room.id === "e2e-node-storage-addon-indexeddb",
-        }),
-    ],
     initialStorage: yjs.doc(() => ({
         messages: yjs.array([
             yjs.object({
