+++ tests/e2e/src/tests/yjs/node-redis/presence/index.test.ts
@@ -3,7 +3,7 @@ import { oneLine } from "common-tags";
 import ms from "ms";
 import { openTestPage, waitMs } from "../../../../utils";
 
-const TEST_URL = "http://localhost:3100/node-redis/presence";
+const TEST_URL = "http://localhost:3100/yjs/node-redis/presence";
 
 test.describe("Node Redis Presence", () => {
     test(
