+++ tests/e2e/src/tests/yjs/node/storage/index.test.ts
@@ -3,7 +3,7 @@ import { oneLine } from "common-tags";
 import ms from "ms";
 import { openTestPage, waitMs } from "../../../../utils";
 
-const TEST_URL = "http://localhost:3100/node/storage";
+const TEST_URL = "http://localhost:3100/yjs/node/storage";
 
 test.describe("Node Storage", () => {
     test(
